+++ lib/settings/widgets/section_graphic_theme.dart
