+++ lib/settings/widgets/section_session.dart
