+++ lib/settings/widgets/settings_partition_controls.dart
@@ -47,8 +47,10 @@ class PartitionControls extends ConsumerWidget {
                 }
               },
               tooltip: '''
+
               The percentage of data allocated for training the model. Ensure
               the total across training, ${useValidation ? "validation" : "tuning"}, and testing sums to 100%.
+
               ''',
             ),
             SettingNumberField(
@@ -62,8 +64,10 @@ class PartitionControls extends ConsumerWidget {
                 }
               },
               tooltip: '''
+
               The percentage of data allocated for ${useValidation ? "validating" : "tuning"} the model. Ensure the total across
               training, ${useValidation ? "validation" : "tuning"}, and testing sums to 100%.
+
               ''',
             ),
             SettingNumberField(
@@ -77,8 +81,10 @@ class PartitionControls extends ConsumerWidget {
                 }
               },
               tooltip: '''
+
               The percentage of data allocated for testing the model. Ensure the total
               across training, ${useValidation ? "validation" : "tuning"}, and testing sums to 100%.
+
               ''',
             ),
             Text(
@@ -92,11 +98,13 @@ class PartitionControls extends ConsumerWidget {
             Container(
               child: MarkdownTooltip(
                 message: '''
+
                 Some data scientists think of the second dataset of the
                 partitions as a dataset to use for **tuning** the model. Others
                 see it as a dataset for **validating** parameter settings. You
                 can choose your preference for the nomenclature here. The choice
                 does not have any material impact on any analysis.
+                
                 ''',
                 child: Row(
                   children: [
